+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2026-04-26
 
+- 修复 macOS 证书安装完成标记复用导致新 root.crt 未被系统信任的问题 @leereliu
 - 修复 macOS 证书安装弹框循环出现导致应用卡住的问题 @leereliu
 - Add GitHub Actions release workflow for macOS dmg packages on `main` branch updates @leereliu
 
@@ -18,4 +19,3 @@
 
 - 修复部分样式错误 @xcodebuild
 - Fix some style errors @xcodebuild
-